refactor(pipeline/network): extract Polling from NetworkDiscovery
Phase 4 commit 2/9 of NetworkDiscovery split (master plan
pipeline-decoupling-master-2026-05-28).

- Carve Polling (FirstIdPredicate, IPollFirstIdArgs, pollTick,
  pollFirstId, awaitFirstId, ITrafficWaitArgs, awaitTraffic, plus the
  internal findTrafficHit helper used by awaitTraffic) from
  NetworkDiscovery.ts (1746 -> 1623 LoC).
- New file:
  src/Scrapers/Pipeline/Mediator/Network/Polling/NetworkPolling.ts
  (~152 LoC).
- NetworkDiscovery.ts now imports awaitFirstId/awaitTraffic and the
  FirstIdPredicate type from the new sub-module; drops the unused
  createPromise / NETWORK_WAIT_FIRST_ID_POLL_MS imports.
- No behavior change; no test BODY change; public API unchanged.

Master plan: pipeline-decoupling-master-2026-05-28 / phase-4
